Add ease in/out balance to "auto" cycler
Adds a 0%-100% slider to the cycler, which moves the point at which the
cycler transitions from easing in, to easing out (ergo, moves the point
of maximum speed during the cycle). This implementation maintains the
current exponential ease in/out curves, but determines their length
dynamically depending on the position of that midpoint.